requirements.txt
django
djangorestframework
django-extensions
ipython